backend/requirements.txt
# Django Backend Requirements
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
django-mptt>=0.14
python-dotenv>=1.0
dj-database-url>=2.1
psycopg2-binary>=2.9
